add functioning delete and find by id methods

+++ src/main/java/cl/omarin/springstudyproject/service/BookService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class BookService {
@@ -25,8 +27,23 @@ public class BookService {
         return bookRepository.save(book);
     }
 
-    public void deleteBook(Long id){
-        bookRepository.deleteById(id);
-        System.out.println("Book Deleted!");
+    public Book deleteBook(Long id){
+        if(bookRepository.findById(id).isPresent()){
+            Book toDelete = bookRepository.findById(id).get();
+            bookRepository.deleteById(id);
+            return toDelete;
+        }
+
+        throw new NoSuchElementException();
+
+    }
+
+    public Book findBook(Long id) {
+        if (bookRepository.findById(id).isPresent()) {
+            return bookRepository.findById(id).get();
+        }else{
+            throw new NoSuchElementException();
+        }
+
     }
 }
